requirements.txt
autopep8==1.4.4
boto==2.49.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
futures==3.3.0
gevent==1.4.0
google-api-core==1.14.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-storage==1.17.0
google-compute-engine==2.8.13
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
greenlet==0.4.15
idna==2.8
Pillow==6.2.1
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
rsa==4.0
signalr-client==0.0.7
six==1.12.0
sseclient==0.0.24
typing==3.7.4.1
urllib3==1.25.3
virtualenv==16.7.1
websocket-client==0.56.0
